Fix bug in backoff resetting
Fixes bug that redefines the backoff on every iteration of slots. 
Initially declares it to None for each node. If the node is ready to transmit AND the backoff is defined as none, calculates the a backoff value.
The backoff value is then reset to None on a succ transmission or a collision (which increases the CW) and forces a new calculation of a backoff value.